pyproject.toml
[tool.poetry]
name = "ghost_blog_automation"
version = "0.1.0"
description = ""
authors = ["Todd Birchard <todd@hackersandslackers.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Pyjwt = "*"
Requests = "*"
Loguru = "*"

[tool.poetry.dev-dependencies]
pytest = "^4.6"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
